chore: Make the binary compile
