+++ pkg/compactor/deletion/delete_requests_manager.go
@@ -21,7 +21,7 @@ import (
 	"github.com/grafana/loki/v3/pkg/compactor/deletionmode"
 	"github.com/grafana/loki/v3/pkg/compactor/retention"
 	"github.com/grafana/loki/v3/pkg/storage/chunk/client"
-	"github.com/grafana/loki/v3/pkg/storage/stores/shipper/indexshipper/util"
+	"github.com/grafana/loki/v3/pkg/storage/chunk/client/local"
 	"github.com/grafana/loki/v3/pkg/util/filter"
 	util_log "github.com/grafana/loki/v3/pkg/util/log"
 )
@@ -134,7 +134,7 @@ func (d *DeleteRequestsManager) Init(tablesManager TablesManager, registerer pro
 		}
 	} else {
 		// Open boltdb file for storing series progress.
-		db, err := util.SafeOpenBoltdbFile(filepath.Join(d.workingDir, seriesProgressFilename))
+		db, err := local.SafeOpenBoltdbFile(filepath.Join(d.workingDir, seriesProgressFilename))
 		if err != nil {
 			return err
 		}
@@ -604,7 +604,7 @@ func (d *DeleteRequestsManager) resetSeriesProgress() {
 		return
 	}
 
-	db, err := util.SafeOpenBoltdbFile(filepath.Join(d.workingDir, seriesProgressFilename))
+	db, err := local.SafeOpenBoltdbFile(filepath.Join(d.workingDir, seriesProgressFilename))
 	if err != nil {
 		level.Error(util_log.Logger).Log("msg", "failed to open series progress", "err", err)
 		return
