+++ pkg/logproto/logproto.proto
@@ -379,28 +379,6 @@ message IndexSeries {
   ];
 }
 
-message QueryIndexResponse {
-  string QueryKey = 1;
-  repeated Row rows = 2;
-}
-
-message Row {
-  bytes rangeValue = 1;
-  bytes value = 2;
-}
-
-message QueryIndexRequest {
-  repeated IndexQuery Queries = 1;
-}
-
-message IndexQuery {
-  string tableName = 1;
-  string hashValue = 2;
-  bytes rangeValuePrefix = 3;
-  bytes rangeValueStart = 4;
-  bytes valueEqual = 5;
-}
-
 message IndexStatsRequest {
   int64 from = 1 [
     (gogoproto.customtype) = "github.com/prometheus/common/model.Time",
