+++ pkg/storage/chunk/client/client.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 
 	"github.com/grafana/loki/v3/pkg/storage/chunk"
-	"github.com/grafana/loki/v3/pkg/storage/stores/series/index"
 )
 
 var (
@@ -23,10 +22,4 @@ type Client interface {
 	DeleteChunk(ctx context.Context, userID, chunkID string) error
 	IsChunkNotFoundErr(err error) bool
 	IsRetryableErr(err error) bool
-}
-
-// ObjectAndIndexClient allows optimisations where the same client handles both
-// Only used by DynamoDB (dynamodbIndexReader and dynamoDBStorageClient)
-type ObjectAndIndexClient interface {
-	PutChunksAndIndex(ctx context.Context, chunks []chunk.Chunk, index index.WriteBatch) error
 }
