+++ pkg/storage/chunk/client/local/boltdb_queries.go
@@ -1,12 +1,11 @@
-package util //nolint:revive
+// Moved a bunch of functions here to keep all BoltDB code local
+// Ideally this would live in its own package.
+package local
 
 import (
 	"context"
 	"sync"
-
-	"github.com/grafana/dskit/concurrency"
-
-	"github.com/grafana/loki/v3/pkg/storage/stores/series/index"
+	"unsafe"
 )
 
 const (
@@ -14,14 +13,14 @@ const (
 	maxConcurrency  = 10
 )
 
-type QueryIndexFunc func(ctx context.Context, queries []index.Query, callback index.QueryPagesCallback) error
+type QueryIndexFunc func(ctx context.Context, queries []Query, callback QueryPagesCallback) error
 
 // QueriesByTable groups and returns queries by tables.
-func QueriesByTable(queries []index.Query) map[string][]index.Query {
-	queriesByTable := make(map[string][]index.Query)
+func QueriesByTable(queries []Query) map[string][]Query {
+	queriesByTable := make(map[string][]Query)
 	for _, query := range queries {
 		if _, ok := queriesByTable[query.TableName]; !ok {
-			queriesByTable[query.TableName] = []index.Query{}
+			queriesByTable[query.TableName] = []Query{}
 		}
 
 		queriesByTable[query.TableName] = append(queriesByTable[query.TableName], query)
@@ -30,32 +29,14 @@ func QueriesByTable(queries []index.Query) map[string][]index.Query {
 	return queriesByTable
 }
 
-func DoParallelQueries(ctx context.Context, queryIndex QueryIndexFunc, queries []index.Query, callback index.QueryPagesCallback) error {
-	if len(queries) == 0 {
-		return nil
-	}
-	if len(queries) <= maxQueriesBatch {
-		return queryIndex(ctx, queries, NewCallbackDeduper(callback, len(queries)))
-	}
-
-	jobsCount := len(queries) / maxQueriesBatch
-	if len(queries)%maxQueriesBatch != 0 {
-		jobsCount++
-	}
-	callback = NewSyncCallbackDeduper(callback, len(queries))
-	return concurrency.ForEachJob(ctx, jobsCount, maxConcurrency, func(ctx context.Context, idx int) error {
-		return queryIndex(ctx, queries[idx*maxQueriesBatch:min((idx+1)*maxQueriesBatch, len(queries))], callback)
-	})
-}
-
 // NewSyncCallbackDeduper should always be used on table level not the whole query level because it just looks at range values which can be repeated across tables
 // NewSyncCallbackDeduper is safe to used by multiple goroutines
 // Cortex anyways dedupes entries across tables
-func NewSyncCallbackDeduper(callback index.QueryPagesCallback, queries int) index.QueryPagesCallback {
+func NewSyncCallbackDeduper(callback QueryPagesCallback, queries int) QueryPagesCallback {
 	syncMap := &syncMap{
 		seen: make(map[string]map[string]struct{}, queries),
 	}
-	return func(q index.Query, rbr index.ReadBatchResult) bool {
+	return func(q Query, rbr ReadBatchResult) bool {
 		return callback(q, &readBatchDeduperSync{
 			syncMap:           syncMap,
 			hashValue:         q.HashValue,
@@ -67,11 +48,11 @@ func NewSyncCallbackDeduper(callback index.QueryPagesCallback, queries int) inde
 // NewCallbackDeduper should always be used on table level not the whole query level because it just looks at range values which can be repeated across tables
 // NewCallbackDeduper is safe not to used by multiple goroutines
 // Cortex anyways dedupes entries across tables
-func NewCallbackDeduper(callback index.QueryPagesCallback, queries int) index.QueryPagesCallback {
+func NewCallbackDeduper(callback QueryPagesCallback, queries int) QueryPagesCallback {
 	f := &readBatchDeduper{
 		seen: make(map[string]map[string]struct{}, queries),
 	}
-	return func(q index.Query, rbr index.ReadBatchResult) bool {
+	return func(q Query, rbr ReadBatchResult) bool {
 		f.hashValue = q.HashValue
 		f.ReadBatchIterator = rbr.Iterator()
 		return callback(q, f)
@@ -79,12 +60,12 @@ func NewCallbackDeduper(callback index.QueryPagesCallback, queries int) index.Qu
 }
 
 type readBatchDeduper struct {
-	index.ReadBatchIterator
+	ReadBatchIterator
 	hashValue string
 	seen      map[string]map[string]struct{}
 }
 
-func (f *readBatchDeduper) Iterator() index.ReadBatchIterator {
+func (f *readBatchDeduper) Iterator() ReadBatchIterator {
 	return f
 }
 
@@ -115,12 +96,12 @@ type syncMap struct {
 }
 
 type readBatchDeduperSync struct {
-	index.ReadBatchIterator
+	ReadBatchIterator
 	hashValue string
 	*syncMap
 }
 
-func (f *readBatchDeduperSync) Iterator() index.ReadBatchIterator {
+func (f *readBatchDeduperSync) Iterator() ReadBatchIterator {
 	return f
 }
 
@@ -159,4 +140,12 @@ func (f *readBatchDeduperSync) Next() bool {
 	}
 
 	return false
+}
+
+func GetUnsafeBytes(s string) []byte {
+	return *((*[]byte)(unsafe.Pointer(&s))) // #nosec G103 -- we know the string is not mutated -- nosemgrep: use-of-unsafe-block
+}
+
+func GetUnsafeString(buf []byte) string {
+	return *((*string)(unsafe.Pointer(&buf))) // #nosec G103 -- we know the string is not mutated -- nosemgrep: use-of-unsafe-block
 }
