+++ pkg/storage/chunk/client/util/util.go
@@ -6,62 +6,7 @@ import (
 	"io"
 	"io/fs"
 	"os"
-
-	"github.com/grafana/loki/v3/pkg/storage/stores/series/index"
 )
-
-// DoSingleQuery is the interface for indexes that don't support batching yet.
-type DoSingleQuery func(context.Context, index.Query, index.QueryPagesCallback) error
-
-// QueryParallelism is the maximum number of subqueries run in
-// parallel per higher-level query
-var QueryParallelism = 100
-
-// DoParallelQueries translates between our interface for query batching,
-// and indexes that don't yet support batching.
-func DoParallelQueries(
-	ctx context.Context, doSingleQuery DoSingleQuery, queries []index.Query,
-	callback index.QueryPagesCallback,
-) error {
-	if len(queries) == 1 {
-		return doSingleQuery(ctx, queries[0], callback)
-	}
-
-	queue := make(chan index.Query)
-	incomingErrors := make(chan error)
-	n := min(len(queries), QueryParallelism)
-	// Run n parallel goroutines fetching queries from the queue
-	for i := 0; i < n; i++ {
-		go func() {
-			ctx, sp := tracer.Start(ctx, "DoParallelQueries-worker")
-			defer sp.End()
-			for {
-				query, ok := <-queue
-				if !ok {
-					return
-				}
-				incomingErrors <- doSingleQuery(ctx, query, callback)
-			}
-		}()
-	}
-	// Send all the queries into the queue
-	go func() {
-		for _, query := range queries {
-			queue <- query
-		}
-		close(queue)
-	}()
-
-	// Now receive all the results.
-	var lastErr error
-	for i := 0; i < len(queries); i++ {
-		err := <-incomingErrors
-		if err != nil {
-			lastErr = err
-		}
-	}
-	return lastErr
-}
 
 // EnsureDirectory makes sure directory is there, if not creates it if not
 func EnsureDirectory(dir string) error {
